requirements.txt
asgiref==3.2.10
Django==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
Pygments==2.7.1
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
